fix(threadx): Comment out VTOR setup which was causing a Hard Fault
A 256bytes padding was added in the beginning of the vector table for
NULL-pointer protection, so the vector table was relocated. See Line 374
in the `tm4c123gh6pm_startup_ccs_gcc.c` file where the VTOR was set. The
VTOR setup inside the `tx_initialize_low_level.S` was overwriting this
value causing the table to point to the wrong interrupt handler
location, hence, the Hard Fault.